Migrate DeepSeek kernels to SFPU macro API.
