--- models/demos/deepseek_v3_b1/kernel_includes/tt_metal/hw/ckernels/blackhole/metal/llk_api/llk_sfpu/llk_math_sdpa_broadcast_col.h
@@ -1,28 +1,0 @@
-// SPDX-FileCopyrightText: © 2025 Tenstorrent USA, Inc.
-//
-// SPDX-License-Identifier: Apache-2.0
-
-#pragma once
-
-#include "llk_math_eltwise_unary_sfpu_init.h"
-#include "llk_math_eltwise_unary_sfpu_params.h"
-#include "ckernel_sfpu_sdpa_broadcast_col.h"
-
-namespace ckernel {
-
-template <bool APPROXIMATE, bool is_fp32_dest_acc_en, DataFormat format, EltwiseBinaryType binary_type>
-inline void llk_math_sfpu_sdpa_broadcast_col_init() {
-    sfpu::init_sdpa_broadcast_col<format, binary_type>();
-}
-
-template <bool APPROXIMATE, bool is_fp32_dest_acc_en, DataFormat format, uint32_t block_width>
-inline void llk_math_sfpu_sdpa_broadcast_mul_col(uint dst_index) {
-    sfpu::calculate_sdpa_broadcast_mul_col<format, block_width>(dst_index);
-}
-
-template <bool APPROXIMATE, bool is_fp32_dest_acc_en, DataFormat format, uint32_t block_width>
-inline void llk_math_sfpu_sdpa_broadcast_sub_col(uint dst_index) {
-    sfpu::calculate_sdpa_broadcast_sub_col<format, block_width>(dst_index);
-}
-
-}  // namespace ckernel
