+++ models/demos/deepseek_v3_b1/kernel_includes/tt_metal/include/compute_kernel_api/sdpa.h
@@ -16,15 +16,17 @@
 #ifdef TRISC_MATH
 #include "../../hw/ckernels/blackhole/metal/llk_api/llk_math_sdpa_bcast_col_srcb_reuse_api.h"
 #include "../../hw/ckernels/blackhole/metal/llk_api/llk_math_sdpa_bcast_col_srca_srcb_reuse_api.h"
-#include "../../hw/ckernels/blackhole/metal/llk_api/llk_sfpu/llk_math_sdpa_reduce_row.h"
+#include "../../hw/ckernels/blackhole/metal/llk_api/llk_sfpu/ckernel_sfpu_sdpa_reduce_row.h"
+#include "../../hw/ckernels/blackhole/metal/llk_api/llk_sfpu/ckernel_sfpu_deepseek_sdpa.h"
 #include "ckernel_sfpu_exp.h"
 #include "ckernel_sfpu_recip.h"
+#include "llk_math_eltwise_unary_sfpu_macros.h"
 #endif
 #ifdef TRISC_UNPACK
 #include "../../hw/ckernels/blackhole/metal/llk_api/llk_unpack_A_sdpa_api.h"
 #endif
 #ifdef TRISC_PACK
-#include "../../../../kernel_includes/tt_metal/hw/ckernels/blackhole/metal/llk_api/llk_sfpu/llk_math_sdpa_reduce_row.h"
+#include "../../../../kernel_includes/tt_metal/hw/ckernels/blackhole/metal/llk_api/llk_sfpu/ckernel_sfpu_sdpa_reduce_row.h"
 #include "ckernel_sfpu_exp.h"
 #include "ckernel_sfpu_recip.h"
 #include "llk_math_eltwise_unary_sfpu_macros.h"
@@ -130,19 +132,35 @@ ALWI void sdpa_mul_bcast_col_srca_srcb_reuse_tiles(uint32_t dst_tile_index) {
 
 template <DataFormat format>
 ALWI void sdpa_reduce_row_init() {
-    MATH((llk_math_sfpu_sdpa_reduce_row_init<APPROX, DST_ACCUM_MODE, format>()));
+    MATH(SFPU_UNARY_INIT_FN(reduce, sfpu::init_sdpa_reduce_row, (format)));
 }
 
 template <DataFormat format, uint32_t block_width>
 ALWI void sdpa_reduce_max_row(uint src_index, uint dst_index, bool prev_max = false) {
-    MATH((llk_math_sfpu_sdpa_reduce_max_row<APPROX, DST_ACCUM_MODE, format, block_width>(
-        src_index, dst_index, prev_max)));
+    MATH(SFPU_UNARY_CALL(
+        DST_SYNC_MODE,
+        DST_ACCUM_MODE,
+        calculate_sdpa_reduce_max_row,
+        (format, block_width),
+        0 /*dst_index*/,
+        VectorMode::None,
+        src_index,
+        dst_index,
+        prev_max));
 }
 
 template <DataFormat format, uint32_t block_width>
 ALWI void sdpa_reduce_sum_row(uint src_index, uint dst_index, bool prev_sum = false) {
-    MATH((llk_math_sfpu_sdpa_reduce_sum_row<APPROX, DST_ACCUM_MODE, format, block_width>(
-        src_index, dst_index, prev_sum)));
+    MATH(SFPU_UNARY_CALL(
+        DST_SYNC_MODE,
+        DST_ACCUM_MODE,
+        calculate_sdpa_reduce_sum_row,
+        (format, block_width),
+        0 /*dst_index*/,
+        VectorMode::None,
+        src_index,
+        dst_index,
+        prev_sum));
 }
 
 #ifdef TRISC_PACK
@@ -282,8 +300,16 @@ void compute_sdpa_chunk(
     sdpa_custom_mm_block<transpose_k>(cb_q, cb_k, cb_mask, 0, 0, mm1_dst_offset, num_tiles_k, chunk_size, mask_chunk);
 
     // Reduce Max (SFPU)
-    PACK((llk_math_sfpu_sdpa_reduce_max_row<false, DST_ACCUM_MODE, DataFormat::Float16_b, chunk_size>(
-        mm1_dst_offset, max_dst_offset, !first_chunk)));
+    PACK(SFPU_UNARY_CALL(
+        DST_SYNC_MODE,
+        DST_ACCUM_MODE,
+        calculate_sdpa_reduce_max_row,
+        (DataFormat::Float16_b, chunk_size),
+        0 /*dst_index*/,
+        VectorMode::None,
+        mm1_dst_offset,
+        max_dst_offset,
+        !first_chunk));
     // Bcast Sub (FPU)
     // Wait for SFPU to finish (sem is 0)
     sdpa_sub_bcast_col_srca_srcb_reuse_tiles_init<chunk_size>(cb_q);  // For tile shape
@@ -336,8 +362,16 @@ void compute_sdpa_chunk(
 
     // Reduce Sum (SFPU)
     PACK((ckernel::sfpu::_init_sdpa_reduce_sum_row_8x32_replay_buffers_()));
-    PACK((llk_math_sfpu_sdpa_reduce_sum_row<false, DST_ACCUM_MODE, DataFormat::Float16_b, chunk_size, true>(
-        mm1_dst_offset, sum_dst_offset, !first_chunk)));
+    PACK(SFPU_UNARY_CALL(
+        DST_SYNC_MODE,
+        DST_ACCUM_MODE,
+        calculate_sdpa_reduce_sum_row,
+        (DataFormat::Float16_b, chunk_size, true /*skip_signalling*/),
+        0 /*dst_index*/,
+        VectorMode::None,
+        mm1_dst_offset,
+        sum_dst_offset,
+        !first_chunk));
     // Signal SFPU is done for the chunk
     if (!first_chunk) {
         // Wait for FPU to signal (this doesn't block SFPU logic)
@@ -367,82 +401,19 @@ void compute_sdpa_recip(uint32_t cb_q, uint32_t sum_dst_offset, uint32_t recip_d
 #ifdef TRISC_MATH
 
 /**
- * The custom SFPI LLK function computes the following operation:
- * cur_max = max(prev_max, worker_max)
- * cur_sum = exp((worker_max - cur_max) * scale) * worker_sum + exp((prev_max - cur_max) * scale) * prev_sum
- * There are 4 results produced:
- * 1. exp_max_diff = exp((worker_max - cur_max) * scale), produced in dst_reg[prev_max_base_idx]
- * 2. exp_max_diff_2 = exp((prev_max - cur_max) * scale), produced in dst_reg[worker_max_base_idx]
- * 3. cur_sum produced in dst_reg[prev_sum_base_idx]
- * 4. cur_max produced in dst_reg[cur_max_base_idx]
- * If final_norm is true, the output is:
- * 1. exp_max_diff = exp((worker_max - cur_max) * scale) * recip(cur_sum), produced in dst_reg[prev_max_base_idx]
- * 2. exp_max_diff_2 = exp((prev_max - cur_max) * scale) * recip(cur_sum), produced in dst_reg[worker_max_base_idx]
- * fused_max_sub_exp_add_tile
- */
-template <bool SDPA_EXP_APPROX_MODE, bool final_norm = false>
-void calculate_fused_max_sub_exp_add_tile(int scale_bf16) {
-    // Non-Approx mode for exp initializes recip for final normalization
-    static_assert(!(final_norm && SDPA_EXP_APPROX_MODE), "Approx mode must be disabled when final_norm is true");
-
-    // 8 rows
-    constexpr int ITERATIONS_HALF_FACE = 2;
-    constexpr uint32_t prev_max_base_idx = 0;     // Tile 0, col 0
-    constexpr uint32_t prev_sum_base_idx = 1;     // Tile 0, col 1
-    constexpr uint32_t worker_max_base_idx = 32;  // Tile 1, col 0
-    constexpr uint32_t worker_sum_base_idx = 33;  // Tile 1, col 1
-    constexpr uint32_t cur_max_base_idx = 64;     // Tile 2, col 0 (output)
-    constexpr uint32_t cur_sum_base_idx = 65;     // Tile 2, col 1 (output)
-
-    for (int d = 0; d < ITERATIONS_HALF_FACE; d++) {
-        // Load inputs for this vector-slot into temporaries to avoid aliasing on dst_reg
-        sfpi::vFloat prev_max_vec = sfpi::dst_reg[prev_max_base_idx];
-        sfpi::vFloat worker_max_vec = sfpi::dst_reg[worker_max_base_idx];
-        sfpi::vFloat prev_sum_vec = sfpi::dst_reg[prev_sum_base_idx];
-        sfpi::vFloat worker_sum_vec = sfpi::dst_reg[worker_sum_base_idx];
-        sfpi::vFloat cur_max;
-        v_if(prev_max_vec < worker_max_vec) { cur_max = worker_max_vec; }
-        v_else { cur_max = prev_max_vec; }
-        v_endif;
-        if constexpr (!final_norm) {
-            sfpi::dst_reg[cur_max_base_idx] = cur_max;
-        }
-
-        // Compute differences
-        sfpi::vFloat diff_prev = prev_max_vec - cur_max;
-        sfpi::vFloat diff_worker = worker_max_vec - cur_max;
-
-        // Exponentials of differences
-        sfpi::vFloat exp_prev =
-            sfpu::_ckernel_sfpu_exp_accurate_<true /*SCALE_EN*/, DST_ACCUM_MODE /*is_fp32_dest_acc_en*/>(
-                diff_prev, scale_bf16);
-        sfpi::vFloat exp_worker =
-            sfpu::_ckernel_sfpu_exp_accurate_<true /*SCALE_EN*/, DST_ACCUM_MODE /*is_fp32_dest_acc_en*/>(
-                diff_worker, scale_bf16);
-
-        if constexpr (!final_norm) {
-            sfpi::dst_reg[cur_sum_base_idx] = exp_worker * worker_sum_vec + exp_prev * prev_sum_vec;
-            sfpi::dst_reg[prev_max_base_idx] = exp_prev;
-            sfpi::dst_reg[worker_max_base_idx] = exp_worker;
-        } else {
-            sfpi::vFloat curr_sum = exp_worker * worker_sum_vec + exp_prev * prev_sum_vec;
-            ckernel::sfpu::_init_sfpu_reciprocal_<false>();
-            sfpi::vFloat recip_sum = ckernel::sfpu::sfpu_reciprocal<SDPA_EXP_APPROX_MODE>(curr_sum);
-            sfpi::dst_reg[prev_max_base_idx] = exp_prev * recip_sum;
-            sfpi::dst_reg[worker_max_base_idx] = exp_worker * recip_sum;
-        }
-        sfpi::dst_reg += 2;
-    }
-}
-
-/**
  * Wrapper for fused max-sub-exp-add SFPI kernel.
- * Invokes calculate_fused_max_sub_exp_add_tile via LLK unary SFPU parameters.
+ * Invokes calculate_fused_max_sub_exp_add_tile via the SFPU macro wrapper.
  */
 template <bool SDPA_EXP_APPROX_MODE, VectorMode vector_mode = VectorMode::C, bool final_norm = false>
 void fused_max_sub_exp_add_tile(uint32_t idst, int scale_bf16) {
-    _llk_math_eltwise_unary_sfpu_params_(
-        calculate_fused_max_sub_exp_add_tile<SDPA_EXP_APPROX_MODE, final_norm>, idst, vector_mode, scale_bf16);
+    SFPU_UNARY_CALL(
+        DST_SYNC_MODE,
+        DST_ACCUM_MODE,
+        calculate_fused_max_sub_exp_add_tile,
+        (SDPA_EXP_APPROX_MODE, final_norm),
+        idst,
+        vector_mode,
+        scale_bf16);
 }
 #endif
